data_dev/src/mwaa/requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-amazon
apache-airflow[amazon]
apache-airflow-providers-postgres[amazon]